+++ pennylane/queuing.py
@@ -17,3 +17,11 @@ The previous location for pennylane.core.queuing.
 
 # pylint: disable=wildcard-import, unused-wildcard-import
 from pennylane.core.queuing import *  # tach-ignore
+
+
+def __getattr__(key):
+    if key == "process_queue":
+        raise AttributeError(
+            "pennylane.queuing.process_queue has been moved to qp.tape.qscript.from_queue"
+        )
+    raise AttributeError(f"module 'pennylane.queuing' has no attribute '{key}'")
